api/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
bottle==0.12.17
certifi==2019.9.11
cffi==1.12.3
Click==7.0
cryptography==2.7
Django==2.1.7
dnspython==1.16.0
importlib-metadata==0.22
more-itertools==7.2.0
pipenv==2018.11.26
pluggy==0.13.0
py==1.8.0
pycparser==2.19
pyOpenSSL==19.0.0
pytest==3.10.1
pytz==2018.9
six==1.12.0
toml==0.10.0
virtualenv==16.4.3
virtualenv-clone==0.5.1
zipp==0.6.0
